Add files via upload
A basic calculator is created to perform basic basic mathematical operations.
This calculator is created using HTML, CSS and JAVASCRIPT, and the basic operations like addition, subtraction, multiplication and division can be performed by using this calculator. This calculator can display both input as well as output in the display screen.